feat(refs DPLAN-12969): Add filters to institution list (4201)
* feat(refs DPLAN-12969): WIP

- Add FilterFlyout to institution list
- Move segments logic from FilterFlyout to parent component

* feat(refs DPLAN-12969, #AB21841): introduce store for FilterFlyout

to be able to easily access filter options in the FilterFlyout component,
they are moved to a store instead of passing them as props. that way,
when loading the grouped or ungrouped filter options after a user
opens a filter dropdown, we don't need to mutate the props, and it's
easier to understand what's going on. grouped and ungrouped
filter options can be written into the store from outside, then be
accessed directly in the FilterFlyout component. we should be able
to remove the data handling inside FilterFlyout now.

* feat(refs DPLAN-12969, #AB21841): get ungrouped options from store

- to be able to get the ungrouped filter options of each filter flyout from
the store, they need to be stored by id in an ungroupedOptions object
- they can then be accessed via the id (of the category which is displayed
as label of the filter flyout)
- to make the store object reactive, we need to use Vue.set (will have to
be removed in vue3, reactivity should work by default)
- since we want to use a flat structure for our data in the FE instead of the
json api structure, the template in FilterFlyoutCheckbox needed to be
adjusted slightly
- isLoading also needs to be handled via the store since each flyout has its
own loading state

* feat(refs DPLAN-12969, #AB21841): load filter options into store

* feat(refs DPLAN-12969,#AB21841): Fix filter count

- Remove redundant code like itemsObject, because these are handled now in store and after requesting the data
- Add store setters for setting selected in group and ungrouped options

* feat(refs DPLAN-12969, #AB21841): allow setting showCount for grouped and ungrouped options

* feat(refs DPLAN-12969,#AB21841): Fix filter search

- Add searched grouped and ungrouped options

* feat(refs DPLAN-12969,#AB21841): Fix filter request

* feat(refs DPLAN-12969,#AB21841): Fix request with initial query

- Store initial filter ids in store
- Watch initial filter ids in Filter Flyout and set queries

* feat(refs DPLAN-12969, #AB21841): allow setting and resetting of filters

- store is updated in FilterFlyout.vue
- filter parameter is added to request only if filterQuery in the store is
not empty
- resetting works by calling the reset method in FilterFlyout.vue from
outside, i.e. in InstitutionList.vue - the store is then updated in said
method

* feat(refs DPLAN-12969,#AB21841): Add some styling

Also order attributes.

* feat(refs DPLAN-12969,#AB21841): Fix loading state in segments list

* feat(refs DPLAN-12969,#AB21841): Remove store for segment filter

We have a new store for filter flyout which handles storing the filter queries and don't need this one anymore.

* style(refs DPLAN-12969, #AB21841): display search and filters inline

* feat(refs DPLAN-12969,#AB21841): Fix filter request for segments list

* feat(refs DPLAN-12969,#AB21841): Add isExpanded to store

Handle isExpanded state in store, so we can check after filter request and focus the input.

* feat(refs DPLAN-12969,#AB21841): Fix syntax error

* style(refs DPLAN-12969, #AB21841): improve spacing

* fix(refs DPLAN-12969, #AB21841): remove non-existing include

InvitableInstitutions don't have a category relationship, and the
categories for the tags are retrieved via assignedTags.category

* style(refs DPLAN-12969, #AB21841): fix layout for small screens

* style(refs DPLAN-12969, #AB21841): rename prop

it has been renamed in demosplan-ui because in other cases
like this, we use variant, too

* style(refs DPLAN-12969, #AB21841): remove unused code

* docs(refs DPLAN-12969, #AB21841): update changelog

* style(refs DPLAN-12969, #AB21841): fix linting errors

* build(refs DPLAN-12969, #AB21841): update yarn.lock after merge

* feat(refs DPLAN-12969, #AB21841): don't overwrite filter param containing search term

* feat(refs DPLAN-12969, #AB21841): remove unused placeholder

the prop in DpSearchField is actually not used, and we can go with
the placeholder set directly in DpSearchField here

* style(refs DPLAN-12969, #AB21841): remove empty space

* style(refs DPLAN-12969, #AB21841): rename event according to conventions

* feat(refs DPLAN-12498, #AB21841): make styling consistent

we no longer want to use a 'filter' label for the filter flyouts
and only the triggers should have a grey background

* feat(refs DPLAN-12498, #AB21841): use correct type

---------

Co-authored-by: Heidrun Wiemold <wiemold@demos-deutschland.de>