fix(refs DPLAN-14736): reduce number of requests (4310)
* fix(refs DPLAN-14736): reduce number of requests

- in the mein berlin addon additional field, after requesting the
dropdown options once, there is not really any need to request
them again without reloading the page
- when filtering the organisation list, the list items are destroyed
and mounted again, so the request was fired 10 times (for each
list item once) with each click on a filter checkbox, leading to a
'too many requests' error when changing the filters a lot
- now, after requesting the dropdown options in the addon
component once, they are emitted to the organisation list component
which is not re-mounted when filtering. the organisation list component
then passes them down via prop back to the addon component. if
available, the request is not fired again. that way, we still have 10
requests initially (which is not ideal, and we might want to find a
solution for that), but no more requests when filtering

* fix(refs DPLAN-14736): add prop to second instance of component

DpOrganisationListItem is used twice here, we should pass the prop in
both instances