fix(refs dplan-14909): Always send assignee and place when setting one.
For some reason, sometimes, when both assignee and place changed, only place where send. Because it happens soemtimes, I think it is some sort of race condition between the request and an update of the initial segment. Because I monitored that the initial segment gets updated aswell even though, It should not at that very moment.

To prevent all that, I decided to bypass the vuex-json-api and update the resourcetype "manually".