fix(refs DPLAN-15247): Always fetch places and assignableUsers for segments once and just once (#4466)
Before, the places and assignableUsers where fetch for every segment.
To prevent these mass of unneccesary request it got checkt if there
are already places/users in the store. If so, don't fire the request.
But if the place or assignee was already set, it was delivered as
relationship with the segment. With that the list wasnt empty anymore.
Because of that and because the if logic wasn't the most elegant way, I moved the logic to the wrapper. there we don't have to check if it was already laoded. we justs fetch on mounted.

* fix(refs DPLAN-15247): Remoce unneccessary and dead code

* fix(refs DPLAN-15247): Remoce unneccessary and dead code