fix(DPLAN-15284): splitStatement: dropdown menu open direction (#4594)
fix(DPLAN-15284): Vue-Multiselect automatically positions the dropdown where there is the most space. The multiselect is inside a scrollable container, and sometimes this causes the dropdown to appear above; because the dropdown menu is absolutely positioned, the container does not scroll it, but simply clips it. Since the dropdowns max-height is 300 px, open it downward for the first six items in the list.