feat_(refs:dplan-12794): FE show organisation drafts (4616)
This feature implements a tabbed interface within the drafts-folder to improve the visibility and organization of DraftStatements. Users can now easily switch between draft statements that belong to the whole organization and their old private drafts. All drafts related to the organization are displayed under the "Organization Drafts" tab, while previous private drafts are shown in the "Private Drafts" tab.

---------

Co-authored-by: Moritz Mandler <mandler@demos-deutschland.de>
Co-authored-by: Demos-CI <devops-system+github@demos-deutschland.de>
Co-authored-by: MoritzMandler <89914798+MoritzMandler@users.noreply.github.com>
Co-authored-by: riechedemos <carolin@demos-international.com>
Co-authored-by: Alex <57663334+martendemos@users.noreply.github.com>