feat (refs DPLAN-15423): WIP: Indicator for change-entry of a custom field is needed to allow using whitelist of fields to track and display in UI.
Because the names of the custom fields are chosen by user, we are can't just list the fields to track in the configuration file (entity_content_change_fields_mapping.yml).
Instead we only can allow-list all or none custom fields. To determine if a change-entry, a change of a custom field on display, we need the flag on the change-entry.