chore(refs dplan-15761): Reduce root emit in layerlist (4726)
in order to move further to real vue3 we have to get rid of root emits.
To archive that, I had to move the logic from a event driven to a state driven approach. That makes the PR a bit larger and maybe not so easy to review. I am sorry for that but I wasn't able to do it differently