feat(refs DPLAN-16008): handle save via custom request
- instead of using the vuex json api lib, we do a custom request
now so that we can send all customField options instead of just
the changed ones to the BE (vuex json api lib only sends the diff,
i.e. the changed options). this is needed because later, when
deleting options, we will not send the deleted options, and with
this approach, all unchanged options would be deleted if we only
send the changed ones