feat(refs BMDS-1551, DPLAN-16430): Allow project-overrides for client bundles (5131)
Client bundles can now be overridden in project configurations in much the same way Twig overrides work. Essentially,
there needs to be a <project>/client/js/bundle/<originalBundleName>.js file and if that exists it will be used instead of the core one.

To access components in the project's client/js/components directory, instead of @DpJs @DpJsProject can be used as an alias for imports.