chore: use flat config for eslint (5102)
* chore: WIP Use flat option for esLint

since esLint 9 they restructured the config file to a "flat" way.

And eslintrc.js is replaced by eslint.config.js

* chore: remove esLint in webpack.config

since long it doesn't help us to find linting flaws. So we decided to move the linting check into a webpack hook.

* chore: Make esLint config in flat-style work

* chore: optimize eslint

* chore: Run lint fix (4118)

* chore: run lint --fix

* fix: Simple esLint errors

* fix: Simple esLint errors

* chore: Update yarn.lock after merge

* chore: update yarn.lock after merge

* chore: add addonDev directory to ignores

this change was done in main in eslintrc, which has been
replaced with eslint.config.js, so it should be added
there now

* chore: remove duplicate import

* refactor: replace eslint-config-standard

- it had not been re-implemented in the flat config, but the
package was still installed
- the package is now removed as it is no longer maintained
- it is replaced with @eslint/js, which is an eslint package
- to avoid a lot of errors (esp. 'x is undefined', some globals
had to be added to the config
- a few more errors were fixed along the way

* chore: adjust event name rules

- since a lot of our event names do not yet follow the
recommended formatting (kebab-case), we need to disable
the 'vue/v-on-event-hyphenation' rule for now
- after renaming our events we can enable it again

* fix: make eslint import-resolver work

- incorrect imports (e.g. of non-existing files) now lead to an eslint
error
- imports from @DpJs and node_modules should be resolved properly
- an import in Statement.spec.js was fixed

* chore: add jest plugin for linting test files

* chore: remove unused eslint dependencies

some are obsolete and others may be added again if the
need arises

* chore: add eslint jquery plugin

running yarn lint . will now also show warnings for jquery
(all rules have been set to warning), including suggestions
for replacing jquery with plain javascript

* chore: make globals globally available

(not just in the .js section)

* chore: remove obsolete eslintrc file

* fix: remove double square brackets

* chore: update yarn.lock after merge

* chore: delete empty file

* chore: add style rules

- when removing eslint-config-standard, these got lost
- we decided in the team that we want to use trailing
commas (not implemented before)

* style: fix eslint errors + warnings (5104)

* style: autofix eslint errors

- after adjusting the eslint rules, we can autofix
a lot of errors and warnings
- most of them seem to be related to vue/attributes-order

* test: update snapshot

* style: autofix linting errors

mostly missing trailing commas, since 'comma-dangle': ['error', 'always-multiline']
was newly introduced

* style: add missing trailing comma

* fix: use buttons for interactive elements (5105)

style: use buttons for interactive elements

- divs should not be used as buttons (and the eslint vuejs-accessibility
plugin throws an error), so they were replaced with actual buttons
- fixes the linting error: visible, non-interactive elements should not
have an interactive handler

* style: autofix linting errors after merge

---------

Co-authored-by: Florian Salis <salis@demos-deutschland.de>
Co-authored-by: salisdemos <40487461+salisdemos@users.noreply.github.com>