style(refs: DPLAN:15522): adjust custom event names (4603)
* style(refs: DPLAN:15522): correct custom event names

Some event names did not follow the naming guidelines.

* style(refs: DPLAN:15522): fix typo in SegmentsList.vue

* chore: autofix linting errors after merge

* chore: fix event name after merge

* style(refs DPLAN-15522): clean up event listeners

- when listening to events, we want to use kebab-case, since this
is needed in the twig files for the events to work, and it's
the format that's officially recommended
(see https://vuejs.org/guide/components/events.html)
- only when emitting events do we use camelCase
- DpInsertableRecommendation is removed as it is only used in
Addons (and duplicated there)
- the event in DpEditField.vue is renamed to 'saveSuccess' again;
it had been renamed, which broke the usage in DpUserListExtendItem.vue.
Now, DpAssessmentTableCard.vue also emits the 'saveSucccess' event, so
nothing should be broken. (It's a temporary solution, because we
need to remove $root.$emits for vue 3)
- slidebar events are kept kebab-case for now, as they are defined in
demosplan-ui in the Slidebar component (so they would have to be changed there, too),
and instead of fixing their naming we should replace them with a better solution

---------

Co-authored-by: Heidrun Wiemold <wiemold@demos-deutschland.de>
Co-authored-by: hwiem <40487668+hwiem@users.noreply.github.com>