feat(refs DPLAN-16555): Add logging output from dplan:frontend:buildinfo (5222)
* feat(refs DPLAN-16555): Add logging output from dplan:frontend:buildinfo

By passing the log output to stderr, we can provide the CSS
counterpart to the JS bundle override listing. This will be
useful for having constant build reminders in adjusted projects that
things aren't quite the way they should be when reasoning from the core.

* chore(refs DPLAN-16555): Restore error printing

A previous refactoring reduced the printed error messages to
[Object(object)].

* style: Apply php-cs-fixer

---------

Co-authored-by: Demos-CI <devops-system+github@demos-deutschland.de>