refactor: Disable extra permission checks
The excel importer is only reachable via feature flags from controllers
where that is already in a user permission check.

It is assumed that the calling service or controller shall perform
the necessary feature checks to keep the importer focused on import tasks.

Additionally, this aligns the segment importer's behaviour with the
statement importer which already doesn't do extra checks.