refactor(refs DPLAN-15761, #AB27466): replace root events (5318)
refactor(refs DPLAN-15761): replace root events

- since root events are no longer supported in vue 3, we need
to replace them
- 'organisation:reset' and 'user:reset' were root-emitted from the
parent, then listened to on the child, which is not how communication
between parent and child should be done
- the reset button is located in the parent, so to reset the local
organisation/user state in the child (OrganisationFormFields/UserFormFields),
a watcher was added to the organisation/user prop, which comes from the
vuex json api store and is reset in the parent on reset. so whenever the
organisation/user is reset in the parent, the local state is updated
in the child
- in addition, reset can be triggered from the CreateItem component, in
which case there is no organisation/user prop. this is solved by passing
a 'triggerReset' prop to the child as well as a watcher, and on reset
in the parent it is set to true, so the local state in the child is reset
and an event emitted that then resets 'shouldResetForm' to false in the
parent (otherwise the watcher would not trigger again on the next reset)