style: display hint in info box
- this way, the interface looks the same in both create and
edit form
- we considered a few options: plain text, warning, and
contextual help icon; since the info is needed only in rare cases
because normally, users will not want to change the platform
procedure template all the time, displaying it as a warning
seemed too obtrusive. displaying it as a tooltip/contextual help
on the other hand seemed to hide the info too much. in the end
we decided to display the text in an info box as that seems less
obtrusive but still draws enough attention when needed