Feat(DPLAN-17020): adds a tags filter to the export modal (#5566)
* DPLAN-16902: This PR adds a tags filter to the export modal and makes it possible to select/filter statements by tags. We already have this functionality in the SegmentsList component, so the FilterFlyout component was taken from there and adjusted to make it more customizable.

to fetch the filter options (filter-options:request), the function sendFilterOptionsRequest was taken from the SegmentsList component and refactored to improve readability (it was very hard to follow).
(SegmentsList: sendFilterOptionsRequest → StatementExportModal: loadFilterFlyoutOptions)

FilterFlyout options are loaded only if no filters are currently active (when the options are not available), which is determined by checking the length of currentQuery

Difference from the design template (was discussed with the PO): it is technically not possible (or very tricky) to render the flyout in front of the modal (as in the design template), so the flyout opens inside the modal window, causing a scrollbar to appear on the right side. To make this behavior more intuitive for users, the function scrollModalToBottom was added so the flyout is focused when opened

FilterFlyout → flyoutHeightStyle was also refactored for better readability, and a check getParentScrollTop was added. When the flyout is placed inside the modal, we must calculate the correct scrollTop value based on the modal instead of the document.

update/add some unit tests for the StatementExportModal component
---------

Co-authored-by: Moritz Mandler <mandler@demos-deutschland.de>
Co-authored-by: Demos-CI <devops-system+github@demos-deutschland.de>
Co-authored-by: bakridemos <102021913+bakridemos@users.noreply.github.com>
Co-authored-by: MoritzMandler <89914798+MoritzMandler@users.noreply.github.com>