fix(refs DPLAN-17118): display legend images for base layers, too
- before, this was only done for overlay layers
- in public detail, in the layer list area, legends are now displayed
in this order: first the uploaded "Planzeichenerklärung" (if available),
second the overlay layer legend files and images, and finally the base layer
legend files and images. this order seems to make sense because if images
are rather high, the downloadable legends are not initially visible and might
be missed by the user. at the same time, legends for the overlays seem to
be more important, since they contain specific information related to the
procedure, so they are displayed above the base layers. also, this follows
the order the layers have in the layer list above.
- styling is adjusted to add some structure to the files and images displayed
in the legend section in the layer list sidebar
- to not show the "Grundkarten" heading in the legend section if no legend
files or images are available for base layers (sometimes, urls for legend images
are broken), we need to check if the provided images were loaded, and only then
do we display the heading
- hover styles are removed for legend images, but the name of the corresponding
layer is displayed on hover
- root event 'layer:toggleLegend' could be removed entirely because we now get
the visibility state from the store