fix(refs: DPLAN-15505): remove additional api call
The getMapOptions after the patch request is unnecessary because the UI is either already updated through local state bindings or we moved away from this view. In the second case, the api call triggered through getMapOptions returns an error because the request hasn't enough time to complete before the router navigates away.